chore: remove z.lazy where possible
