+++ packages/openapi-types/src/schemas/3.1/processed/path-item-object-without-callbacks.ts
@@ -1,5 +1,6 @@
-import { z } from 'zod'
-import { BasePathItemObjectSchema, OperationObjectSchemaWithoutCallbacks } from './shared-path-schemas'
+import type { z } from 'zod'
+import { OperationObjectSchemaWithoutCallbacks } from './operation-object-without-callbacks'
+import { BasePathItemObjectSchema } from './shared-path-schemas'
 
 type PathItemObject = z.infer<typeof BasePathItemObjectSchema> & {
   get?: z.infer<typeof OperationObjectSchemaWithoutCallbacks>
@@ -21,39 +22,37 @@ type PathItemObject = z.infer<typeof BasePathItemObjectSchema> & {
  *
  * @see https://github.com/OAI/OpenAPI-Specification/blob/main/versions/3.1.1.md#path-item-object
  */
-export const PathItemObjectSchemaWithoutCallbacks: z.ZodType<PathItemObject> = z.lazy(() =>
-  BasePathItemObjectSchema.extend({
-    /**
-     * A definition of a GET operation on this path.
-     */
-    get: OperationObjectSchemaWithoutCallbacks.optional(),
-    /**
-     * A definition of a PUT operation on this path.
-     */
-    put: OperationObjectSchemaWithoutCallbacks.optional(),
-    /**
-     * A definition of a POST operation on this path.
-     */
-    post: OperationObjectSchemaWithoutCallbacks.optional(),
-    /**
-     * A definition of a DELETE operation on this path.
-     */
-    delete: OperationObjectSchemaWithoutCallbacks.optional(),
-    /**
-     * A definition of a OPTIONS operation on this path.
-     */
-    options: OperationObjectSchemaWithoutCallbacks.optional(),
-    /**
-     * A definition of a HEAD operation on this path.
-     */
-    head: OperationObjectSchemaWithoutCallbacks.optional(),
-    /**
-     * A definition of a PATCH operation on this path.
-     */
-    patch: OperationObjectSchemaWithoutCallbacks.optional(),
-    /**
-     * A definition of a TRACE operation on this path.
-     */
-    trace: OperationObjectSchemaWithoutCallbacks.optional(),
-  }),
-)
+export const PathItemObjectSchemaWithoutCallbacks: z.ZodType<PathItemObject> = BasePathItemObjectSchema.extend({
+  /**
+   * A definition of a GET operation on this path.
+   */
+  get: OperationObjectSchemaWithoutCallbacks.optional(),
+  /**
+   * A definition of a PUT operation on this path.
+   */
+  put: OperationObjectSchemaWithoutCallbacks.optional(),
+  /**
+   * A definition of a POST operation on this path.
+   */
+  post: OperationObjectSchemaWithoutCallbacks.optional(),
+  /**
+   * A definition of a DELETE operation on this path.
+   */
+  delete: OperationObjectSchemaWithoutCallbacks.optional(),
+  /**
+   * A definition of a OPTIONS operation on this path.
+   */
+  options: OperationObjectSchemaWithoutCallbacks.optional(),
+  /**
+   * A definition of a HEAD operation on this path.
+   */
+  head: OperationObjectSchemaWithoutCallbacks.optional(),
+  /**
+   * A definition of a PATCH operation on this path.
+   */
+  patch: OperationObjectSchemaWithoutCallbacks.optional(),
+  /**
+   * A definition of a TRACE operation on this path.
+   */
+  trace: OperationObjectSchemaWithoutCallbacks.optional(),
+})
