+++ packages/openapi-types/src/schemas/3.1/processed/path-item-object.ts
@@ -1,4 +1,3 @@
-import { z } from 'zod'
 import { OperationObjectSchema } from './operation-object'
 import { BasePathItemObjectSchema } from './shared-path-schemas'
 
@@ -11,39 +10,37 @@ import { BasePathItemObjectSchema } from './shared-path-schemas'
  *
  * @see https://github.com/OAI/OpenAPI-Specification/blob/main/versions/3.1.1.md#path-item-object
  */
-export const PathItemObjectSchema = z.lazy(() =>
-  BasePathItemObjectSchema.extend({
-    /**
-     * A definition of a GET operation on this path.
-     */
-    get: OperationObjectSchema.optional(),
-    /**
-     * A definition of a PUT operation on this path.
-     */
-    put: OperationObjectSchema.optional(),
-    /**
-     * A definition of a POST operation on this path.
-     */
-    post: OperationObjectSchema.optional(),
-    /**
-     * A definition of a DELETE operation on this path.
-     */
-    delete: OperationObjectSchema.optional(),
-    /**
-     * A definition of a OPTIONS operation on this path.
-     */
-    options: OperationObjectSchema.optional(),
-    /**
-     * A definition of a HEAD operation on this path.
-     */
-    head: OperationObjectSchema.optional(),
-    /**
-     * A definition of a PATCH operation on this path.
-     */
-    patch: OperationObjectSchema.optional(),
-    /**
-     * A definition of a TRACE operation on this path.
-     */
-    trace: OperationObjectSchema.optional(),
-  }),
-)
+export const PathItemObjectSchema = BasePathItemObjectSchema.extend({
+  /**
+   * A definition of a GET operation on this path.
+   */
+  get: OperationObjectSchema.optional(),
+  /**
+   * A definition of a PUT operation on this path.
+   */
+  put: OperationObjectSchema.optional(),
+  /**
+   * A definition of a POST operation on this path.
+   */
+  post: OperationObjectSchema.optional(),
+  /**
+   * A definition of a DELETE operation on this path.
+   */
+  delete: OperationObjectSchema.optional(),
+  /**
+   * A definition of a OPTIONS operation on this path.
+   */
+  options: OperationObjectSchema.optional(),
+  /**
+   * A definition of a HEAD operation on this path.
+   */
+  head: OperationObjectSchema.optional(),
+  /**
+   * A definition of a PATCH operation on this path.
+   */
+  patch: OperationObjectSchema.optional(),
+  /**
+   * A definition of a TRACE operation on this path.
+   */
+  trace: OperationObjectSchema.optional(),
+})
